refactor: #3 SignIn DTO 수정
