--- src/main/java/com/tm/gogo/web/auth/SignInDto.java
@@ -1,23 +1,0 @@
-package com.tm.gogo.web.auth;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
-
-public class SignInDto {
-
-    @Getter
-    @NoArgsConstructor
-    @AllArgsConstructor
-    @Builder
-    public static class Request{
-        private String email;
-        private String password;
-
-        public UsernamePasswordAuthenticationToken toAuthentication() {
-            return new UsernamePasswordAuthenticationToken(email, password);
-        }
-    }
-}
